Week2/Ex/requirements.txt
Flask
Flask-Cors
PyJWT